Reorganzied how the components within the user settings.
